Implement mean free path calculation
To determine if the cylinder code is working correctly, we need to
ensure that the mean free path of a nucleon is reproduced. To test this,
a method of determining the mean free path is implemented. This is done
through fixing the density and cross-section inside the nucleus and
propogating particles from the center outwards. Upon interaction, the
distance from the origin is recorded. This allows an extraction of a mean
free path after many events.

Currently, the extracted value and the theory predicted value from the
density and cross-section are not always consistent. This may be due to
the fact that there are multiple nucleons wihtin a cylinder, but still
needs to be verified.