Update settings passing
Due to the initialization of the nucleus, it was not possible to import
the settings code into the nucleus module. This prevented allowing a run
card to set the parameters for the mean free path calculation. The
initialization of the nucleus is thus moved to the main code when the
run card is first loaded.