Implement check for all particles in cylinder
There was an issue with the cylinder if multiple nucleons were inside
the cylinder. This is not corrected by looping over all the
nucleons inside to see if any would interact. There are however still a
few issues that need to be addressed:

1. What if multiple nucleons should be interacted with?
2. What order should we look through the nucleons to choose for
interactions if we only take one at a time?