Update curve fitting for mean free path
Previously, the curve fitting did not give an uncertainty on the
fit. This is now updated to all for a direct comparison between the
true value and that extracted from the fit. Additionally, the
unceratainty in each histogram bin is included to ensure an
accurate estimate of the uncertainty and the true fit value.