Implement plane interaction check and calculation
Check if an interaction is allowed between the propagating nucleon and
all other nucleons. This check also gets the distance squared of closest
approach between the propagating nucleon and each allowed nucleon.
Additionally, calculate the probability of interacting with each
nucleon. Use a sorted list of distances to determine probabilistically
which nucleons interacted if any.